* Link against libfontconfig when required by libXft * Various fixes for warnings, misapplied initialisations and missing error   checks in core and VM functions
-- Christoph

+++ src/include/sci_memory.h
@@ -114,7 +114,7 @@
 #define INFO_MEMORY(alloc_statement, size, filename, linenum, funcname)\
 	fprintf(stderr, "ALLOC_MEM(%lu bytes) [%s (%s) : %u] "\
 		#alloc_statement "\n",\
-		size, filename, funcname, linenum);
+		(unsigned long) size, filename, funcname, linenum);
 
 
 #ifdef UNCHECKED_MALLOCS
